refactor(devtools): isolate the developer tools behind one entry point
The developer tools were encapsulated in their own modules, but the
framework still knew about them in three places: Component.js owned the
Ctrl+F12 listener and the lazy control creation, AppState.js carried a
`developerTools` field in its state inventory, and ErrorView.js required
the DeveloperTools module by name to run its Details action. The dialog
itself also still sat in core/ next to the framework modules.

Now:

  core/devtools/  holds the WHOLE feature - DevTools.js (lifecycle
  facade), DeveloperTools.js + .fragment.xml (the dialog, moved here
  from core/), Recorder.js, Inspect.js, Picker.js, LiveEdit.js.

  The framework's entire coupling is two calls:
      Component.init()  ->  DevTools.install()
      Component.exit()  ->  DevTools.exit()

  DevTools.js owns what Component.js used to: the Ctrl+F12 shortcut, the
  lazy dialog creation and teardown, the recorder's install/uninstall,
  and the "?z2ui5-devtools=" auto open.

  AppState.state.developerTools is gone. The instance lives in the
  facade and is still published on the z2ui5 global under its old name,
  so apps that reached it there (via the js_loader popup) keep working.

  The ErrorView dependency is INVERTED. The overlay no longer names any
  developer-tools module: it runs the generic `onErrorDetails` callback
  array (a new, documented AppState extension point) and leaves the
  Details button out entirely when nothing registered - so deleting
  core/devtools/ degrades the overlay gracefully instead of breaking it.
  DevTools.js registers itself there.

  No framework module names a developer-tools module any more; the only
  remaining mentions are comments explaining why some state is kept.

Alongside that, three improvements to the tools themselves:

  - Help tab: what every tab answers, the entry points that are not
    guessable (Ctrl+F12, ?z2ui5-devtools=), what the footer actions do,
    and how to produce a bug report. Discoverability was the real
    barrier - a tab nobody knows about helps nobody.
  - "Copy Tab" puts the current tab's content on the clipboard. A
    CodeEditor has no select-all affordance, so copying one report meant
    dragging across thousands of lines or going through Export.
  - The History tab gains a summary: average and slowest backend time
    with the event that caused it, total and largest response, and the
    count of roundtrips that never rendered. Spotting one slow event in
    50 rows by eye does not work.

Covered by the new devtoolsFacade.spec.js (14 specs: install
idempotence, lazy creation, shortcut, auto open, error-details provider,
teardown and re-install) plus the reworked ErrorView specs for the
inverted dependency - 685 JS specs.
