+++ .github/app2abap/trans2abap.js
@@ -108,8 +108,8 @@ const CLASS_NAME_STEMS = {
     'cc/UITableExt.js': 'uitable_js',
     'cc/UploadSetExt.js': 'upldset_js',
     'cc/Websocket.js': 'websock_js',
-    'core/DeveloperTools.fragment.xml': 'dtools_xml',
-    'core/DeveloperTools.js': 'dtools_js',
+    'core/devtools/DeveloperTools.fragment.xml': 'dtools_xml',
+    'core/devtools/DeveloperTools.js': 'dtools_js',
     'core/ErrorView.js': 'errview_js',
     'core/FrontendAction.js': 'frontact_js',
     'core/ScrollFocus.js': 'scrfocus_js',
